Weekly merge 9 6 2022

+++ src/mpid/ch4/shm/posix/release_gather/nb_release_gather.c
@@ -73,7 +73,7 @@ int MPIDI_POSIX_nb_release_gather_comm_init(MPIR_Comm * comm_ptr,
             ibcast_flags_shm_size = MPIDI_POSIX_RELEASE_GATHER_FLAG_SPACE_PER_RANK * num_ranks *
                 MPIR_CVAR_BCAST_INTRANODE_NUM_CELLS;
             /* Reset flags_shm_size so that the data buffers are aligned to the system pages */
-            ibcast_flags_num_pages = ibcast_flags_shm_size / (int) (pg_sz);
+            ibcast_flags_num_pages = (int) (ibcast_flags_shm_size / pg_sz);
             if (ibcast_flags_shm_size % pg_sz != 0) {
                 ibcast_flags_num_pages++;
             }
@@ -82,7 +82,7 @@ int MPIDI_POSIX_nb_release_gather_comm_init(MPIR_Comm * comm_ptr,
             ireduce_flags_shm_size = MPIDI_POSIX_RELEASE_GATHER_FLAG_SPACE_PER_RANK * num_ranks *
                 MPIR_CVAR_REDUCE_INTRANODE_NUM_CELLS;
             /* Reset flags_shm_size so that the data buffers are aligned to the system pages */
-            ireduce_flags_num_pages = ireduce_flags_shm_size / (int) (pg_sz);
+            ireduce_flags_num_pages = (int) (ireduce_flags_shm_size / pg_sz);
             if (ireduce_flags_shm_size % pg_sz != 0) {
                 ireduce_flags_num_pages++;
             }
@@ -102,7 +102,8 @@ int MPIDI_POSIX_nb_release_gather_comm_init(MPIR_Comm * comm_ptr,
         if (rank == 0) {
             /* rank 0 decides if more memory can be created and broadcasts the decision to other
              * ranks */
-            tmp_shm_counter = MPL_atomic_acquire_load_uint64(MPIDI_POSIX_shm_limit_counter);
+            tmp_shm_counter =
+                (size_t) MPL_atomic_acquire_load_uint64(MPIDI_POSIX_shm_limit_counter);
 
             /* Check if it is allowed to create more shm on this node */
             if ((tmp_shm_counter + memory_to_be_allocated) >
